Parameterized Queries for SQL injection protection

+++ server/controller/habit.controller.js
@@ -7,7 +7,8 @@ const getHabits = async (req, res) => {
     const { username } = req.params;
     const habits = await pool.query(
       `SELECT * FROM "habit" \
-      WHERE username = '${username}';`
+      WHERE username = $1;`,
+      [username]
     );
     res.json(habits.rows);
   } catch {
@@ -19,13 +20,17 @@ const createHabit = async (req, res) => {
   try {
     const { username } = req.params;
     const { name, datestarted, dateended, streakgoal, iconno } = req.body;
-    // TODO: NOT SQL INJECTION SAFE
-    const habits = await pool.query(
-      `INSERT INTO "habit" (name, username, datestarted, dateended, streakgoal, iconno) \
-      VALUES ('${name}', '${username}', TO_DATE('${datestarted}', 'DD/MM/YYYY'), $1, $2, $3)`,
-      [dateended, streakgoal, iconno]
-    );
-    res.json(habits.rows);
+    if (username && name && datestarted && iconno) {
+      // TODO: NOT SQL INJECTION SAFE
+      const habits = await pool.query(
+        `INSERT INTO "habit" (name, username, datestarted, dateended, streakgoal, iconno) \
+        VALUES ($1, $2, TO_DATE($3, 'DD/MM/YYYY'), $4, $5, $6);`,
+        [name, username, datestarted, dateended, streakgoal, iconno]
+      );
+      res.json(habits.rows);
+    } else {
+      res.sendStatus(400);
+    }
   } catch {
     res.sendStatus(400);
   }
@@ -34,12 +39,16 @@ const createHabit = async (req, res) => {
 const deleteHabit = async (req, res) => {
   try {
     const { hid } = req.body;
-    const habits = await pool.query(
-      `DELETE FROM "habit" \
-      WHERE hid=$1`,
-      [hid]
-    );
-    res.json(habits.rows);
+    if (hid) {
+      const habits = await pool.query(
+        `DELETE FROM "habit" \
+        WHERE hid=$1`,
+        [hid]
+      );
+      res.json(habits.rows);
+    } else {
+      res.sendStatus(400);
+    }
   } catch {
     res.sendStatus(400);
   }
@@ -50,12 +59,16 @@ const createEntry = async (req, res) => {
     const { hid } = req.params;
     const { date } = req.body;
     // TODO: if username, email != fall and are both strings else return 4**
-    const entry = await pool.query(
-      `INSERT INTO "entry" (hid, date) \
-      VALUES ($1, TO_DATE('${date}', 'DD/MM/YYYY'))`,
-      [hid]
-    );
-    res.json(entry.rows);
+    if (hid && date) {
+      const entry = await pool.query(
+        `INSERT INTO "entry" (hid, date) \
+        VALUES ($1, TO_DATE($2, 'DD/MM/YYYY'))`,
+        [hid, date]
+      );
+      res.json(entry.rows);
+    } else {
+      res.sendStatus(400);
+    }
   } catch {
     res.sendStatus(400);
   }
@@ -64,13 +77,16 @@ const createEntry = async (req, res) => {
 const deleteEntry = async (req, res) => {
   try {
     const { id } = req.body;
-    console.log(id);
-    const entry = await pool.query(
-      `DELETE FROM "entry" \
-      WHERE id=$1`,
-      [id]
-    );
-    res.json(entry.rows);
+    if (id) {
+      const entry = await pool.query(
+        `DELETE FROM "entry" \
+        WHERE id=$1`,
+        [id]
+      );
+      res.json(entry.rows);
+    } else {
+      res.sendStatus(400);
+    }
   } catch {
     res.sendStatus(400);
   }
@@ -81,12 +97,16 @@ const createNote = async (req, res) => {
     const { hid } = req.params;
     const { note } = req.body;
     // TODO: NOT SQL INJECTION SAFE
-    const notes = await pool.query(
-      `INSERT INTO "notes" (hid, note) \
-      VALUES ($1, '${note}')`,
-      [hid]
-    );
-    res.json(notes.rows);
+    if (hid && note) {
+      const notes = await pool.query(
+        `INSERT INTO "notes" (hid, note) \
+        VALUES ($1, $2)`,
+        [hid, note]
+      );
+      res.json(notes.rows);
+    } else {
+      res.sendStatus(400);
+    }
   } catch {
     res.sendStatus(400);
   }
@@ -95,12 +115,16 @@ const createNote = async (req, res) => {
 const deleteNote = async (req, res) => {
   try {
     const { id } = req.body;
-    const notes = await pool.query(
-      `DELETE FROM "notes" \
-      WHERE id=$1`,
-      [id]
-    );
-    res.json(notes.rows);
+    if (id) {
+      const notes = await pool.query(
+        `DELETE FROM "notes" \
+        WHERE id=$1`,
+        [id]
+      );
+      res.json(notes.rows);
+    } else {
+      res.sendStatus(400);
+    }
   } catch {
     res.sendStatus(400);
   }
@@ -109,12 +133,16 @@ const deleteNote = async (req, res) => {
 const getNotes = async (req, res) => {
   try {
     const { hid } = req.params;
-    const notes = await pool.query(
-      `SELECT * FROM "notes" \
-      WHERE hid=($1)`,
-      [hid]
-    );
-    res.json(notes.rows);
+    if (hid) {
+      const notes = await pool.query(
+        `SELECT * FROM "notes" \
+        WHERE hid=($1)`,
+        [hid]
+      );
+      res.json(notes.rows);
+    } else {
+      res.sendStatus(400);
+    }
   } catch {
     res.sendStatus(400);
   }
